fixed completion rate bug, code still needs to be cleaned up
avg_completion by switching to a double. It was set as an int but the / was creating decimals so the value would return 0.

+++ Tutor.h
@@ -26,7 +26,7 @@ class Tutor: public User{
         void set_days(const std::vector<bool>& days);
 
         double avg_rating();
-        int avg_completion();
+        double avg_completion();
         int get_completed() const;
         int get_matched() const;
         std::vector<bool> get_days();
